Fixing Program Service to work with the new transaction approach.
When creating a new program we are not injecting in the newly created program id. Adding this to the request so that rest of the program is created without error.

Removing manual creating of the manager factory as it is not required.

No need to create MySQLAccounts as it is not necessary.

issue: BMS-1376
reviewer: MatthewB