Updated version of Node, npm, gulp-sass and frontend-maven-plugin.
This is to ensure that our development environment works correctly with our gulp-sass plugin. Currently if a developer has a newer version of node than v0.10.x the gulp-sass plugin will cause issues (due to node-sass and libsass). Upgrading the dependencies fixes these issues and means we can work with a newer version of Node.

Issue: none
Reviewer: Iryna