Upgraded angular from version 1.3.11 to 1.4.3 as the newer version is already being used elsewhere in Workbench.
When upgrading angular we need to keep angular-mocks in sync with the same version.
There was an issue after the upgrade where isolate scope properties that are one way binding (use '@') cannot be assigned a default value if they are not defined on the instance of the directive. This means that a different scope property had to be used, and the one time binding property only used to read from.

Issue: BMS-1176
Reviewer: Iryna