Get variable after update
This allows to get server-side applied logic to the model, like
allowsFormula which is computed base on variable types, and thus
removes business logic from the frontend.
Also makes the view after the update more consistent with the normal
view, which is filled based the GET variable.
Technically PUT shouldn't return an object, but if we relax that
constraint, an alternative would be to return the variable along with
that call.

Issue: BMS-4902 / BMS-5036 / BMS-5040 / BMS-5041
Reviewer: None