Select first available tool for the user
Under the new roles/permission model, manage germplasm may no longer be
available and besides it's not right to have hardcoded there.
To do this we need the key of WorkbenchSidebar.sidebarTreeMap to be the
tool name.
If no tools are available, do nothing.

Issue: IBP-2792 / IBP-1192 / IBP-2912
Reviewer: None