Rewrite authorize screen
The login screen uses css classes and css transitions to hide/show
sections (e.g. recover password). The initial approach for the authorize
window was to use the same form but instead of hiding elements with css
class, use jquery show/hide. For this reason a hidden element was
partially overlapping with the authorize button.
This new approach is to use a completely separate form, better for
separation of concerns. The css transitions is left as an improvement
for the future.

See also the improvement IBP-3605

Issue: IBP-3431